refactor(server): simplify references to utils/file
